Dive_sites-Notes shouldn't be stored as attribute
If so, some formatting chars (like "\n") will be dropped while parsing.
This could be pretty annoying for a user who tries to keep notes other
than a simple text, e.g. if the site is a wreck, may be interesting to
have some data shown like:
			...
		Max. Depth =  60 m
		Min. Depth =  40 m
		Prow: bla, bla, bla.
		Stern: bla, bla, bla.
			...
instead of a single text line.

Signed-off-by: Salvador Cuñat <salvador.cunat@gmail.com>
ACKed-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>