Don't use the same variable name for loop and temp storage
Oops. Moving the check to the beginning of the function in order to avoid
partial execution in commit 69036a1bb777 ("Avoid resource leak by bailing
early") had a nasty side effect. Since Linus used 'i' both to hold the
dive number and as a loop variable, by moving this to the top of the
function the dive nr was overwritten which caused all kinds of problems.

My bad - but of course you shouldn't mix loop variables with meaningful
variables, either...

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>