In free dive mode split dives with only a 10 second surface interval
If the dive computer tells us that this was free diving we should be a lot
more aggressive splitting dives. For scuba dives a minute seems more
reasonable (to avoid splitting dives when you pop to the surface to get
the bearing and continue the dive after a few second).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>