dive.h: don't declare force_fixup_dive() in the header
The compiler complains that said static (not-inline) function
in declared in the header but has no definition.

The function is only used as a helper in dive.c so so this
simply acts as a forward declaration and we can safely remove
it from dive.h. The same forward declaration can be in dive.c.

Since dive.h is include in *many* translation units, this
causes a lot instances of the same warning to be show.

If the function becomes part of the API, the static keyword
needs to be removed and the declaration needs to be added in
the header again.

Signed-off-by: Lubomir I. Ivanov <neolit123@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>