Ensure Out of Source builds
In source builds creates a lot of noise on the source tree,
creating tons of temporaries that makes git status cry, and it also
creates dangerous files that have names similar to actual source
and could make a developer use them by mistake.

Make out of source buiild mandatory to remove that, it also makes
CMake code easier to read.

Signed-off-by: Tomaz Canabrava <tomaz.canabrava@intel.com>